requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
click==6.7
dnspython==1.15.0
Flask==1.0.2
Flask-Cors==3.0.4
gast==0.2.0
grpcio==1.11.0
gunicorn==19.8.1
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.14.3
protobuf==3.5.2.post1
pycodestyle==2.4.0
pydocstyle==2.1.1
pyflakes==2.0.0
pylama==7.4.3
pymongo==3.6.1
six==1.11.0
snowballstemmer==1.2.1
tensorflow==1.8.0
termcolor==1.1.0
Werkzeug==0.14.1
Pillow==5.2.0
h5py==2.8.0
